+++ upload/admin/language/lt-lt/extension/payment/itella_cod.php
@@ -1,9 +1,9 @@
 <?php
 // Heading
-$_['heading_title'] = 'Itella COD';
+$_['heading_title'] = 'Smartposti COD';
 
 // Menu
-$_['menu_itella_cod']               = 'Itella COD';
+$_['menu_itella_cod']               = 'Smartposti COD';
 $_['menu_itella_cod_settings']      = 'COD Nustatymai';
 $_['menu_itella_cod_fee_settings']  = 'Mokesčio Nustatymai';
 
@@ -17,9 +17,9 @@ $_['text_extension']  = 'Išplėtimai';
 $_['text_enabled']    = 'Įjungta';
 $_['text_disabled']   = 'Išjungta';
 $_['text_all_zones']  = '-- Visos --';
-$_['text_itella_cod'] = '<img src="view/image/itella_cod/logo.png" alt="Itella Logo" style="height: 33px;">';
-$_['text_success']		= 'Pavyko: Itella COD nustatymai sėkmingai modifikuoti!';
-$_['text_edit']       = 'Itella COD nustatymai';
+$_['text_itella_cod'] = '<img src="view/image/itella_cod/logo.png" alt="Smartposti Logo" style="height: 33px;">';
+$_['text_success']		= 'Pavyko: Smartposti COD nustatymai sėkmingai modifikuoti!';
+$_['text_edit']       = 'Smartposti COD nustatymai';
 
 // Entry
 $_['entry_total']		      = 'Viso';