+++ upload/catalog/language/lt-lt/extension/total/itella_cod_fee.php
@@ -1,2 +1,2 @@
 <?php
-$_['text_title'] = 'Itella COD Mokestis';
+$_['text_title'] = 'Smartposti COD Mokestis';